# www/html/api-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste Laravel",
        "description": "Api teste Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cep/{cep}": {
            "get": {
                "tags": [
                    "Cep"
                ],
                "summary": "Exibe um cliente",
                "operationId": "a3e55fc6967ddcbbfe7e96b656dba36d",
                "parameters": [
                    {
                        "name": "cep",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do Cep"
                    }
                }
            }
        },
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista os clientes",
                "operationId": "65ff200b3cd0cebf89ab35bd2a55326a",
                "responses": {
                    "200": {
                        "description": "Lista de clientes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastro de clientes",
                "operationId": "650c2927959073663a6078b724eaefbc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "telefone",
                                    "email"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente cadastrado com sucesso"
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Exibe um cliente",
                "operationId": "b54f5bbb2020ce21b50aa66943c3f384",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do cliente"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar informações do cliente",
                "description": "Atualiza as informações de um cliente pelo ID",
                "operationId": "updateCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Informações do cliente atualizadas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "telefone": {
                                            "type": "string",
                                            "example": "1234567890"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "johndoe@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        }
    }
}